pyproject.toml
[tool.poetry]
name = "diss"
version = "0.2.11"
description = "Demonstration Informed Specification Search"
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"  # scipy
attrs = "^22.0"
numpy = "^1.21.2"
scipy = "^1.7.2"
funcy = "^1.16"
jupyterlab = "^3.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-flake8 = "^1.0.7"
pydot = "^1.4.2"
mypy = "^0.910"
dfa = "^4.2"
dfa-identify = "^3.8" 
blessings = "^1.7"
jupyter = "^1.0.0"
matplotlib = "^3.4.3"
seaborn = "^0.11.2"
tqdm = "^4.62.3"
python-sat = "^0.1.7.dev11"
py-aiger = "^6.1.21"
py-aiger-bdd = "^3.1"
py-aiger-coins = "^3.3.6"
py-aiger-gridworld = "^0.4.3"
py-aiger-ptltl = "^3.1.0"
py-aiger-dfa = "^0.4.1"

[tool.poetry.group.dev.dependencies]
aalpy = "^1.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
